+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Major changes:
   * The functions `read_flights()` and `read_airport_movements()` now have a new parameter `cache`, which indicates whether the function should read cached data downloaded previously. Defaults to `TRUE`. Closed #31.
-  * The function `read_aircrafts()` now has a `date` parameter, which allows one to download the data on aircrafts registered at ANAC at particular years/monhts.  Closed #33.
+  * The function `read_aircrafts()` now has a `date` parameter, which allows one to download the data on aircrafts registered at ANAC at particular years/months.  Closed #33.
 
 * Minor changes:
   * All functions now return numeric columns with `numeric` class. Closed #32.
@@ -10,7 +10,8 @@
 * Bug fixes:
   * Fixed bug when unzipping files for `read_flights()` function in Unix systems. 
   Closed #31.
- 
+   * Updated link to private airports data changed by ANAC. Closed #34.
+
 # flightsbr v0.3.0
 
 * Major changes: